autocomplete refactor
