+++ frontend/components/EventCard.tsx
